feat: simple transform utility
We are using LLMs from a research perspective, but it is
relatively straight forward to translate between job specs
by doing a careful mapping. Arguably, there are many different
options for any particular manager, but users only use a subset
and this is not hard to do, and gives us (I think) maybe higher
accuracy. The strategy here is to map each to a standard definition,
and then each transformer knows how to convert the standard to their
spec of choice. For K8s this will get more complex when we want multi-
node and want to choose between flux-operator or trainer, but this
should be a good starting case - a simple job. Using this setup
we can test parsing a lot of jobspecs and understanding 1. the
breakdown of params - what percentage of jobspecs use each one,
and thus which are most important to include. 2. what needs mapping
. For example, for flux to k8s, the main translations are times to
seconds, and then a priority number to a string level. Arguably,
there are always going to be some attributes that are lost. E.g.,
kubernetes has no sense of cpu-affinity per task, so for now
I am being verbose in logging those and adding as labels.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>