flux-validator: validate, count canonical jobspec resources
Add support for validating canonical jobspecs in YAML or JSON format.
The Flux Jobspec class has a function that validates canonical jobspec
and throws errors with specific reasons why an input jobspec is
invalid. Integrate this functionality into the flux-validator.

Also add support for walking a canonical jobspec and validating the
resource counts via `.resource_walk()`. Outputting the counts will
provide feedback for an agent to correct a generated canonical
jobspec.

Update the Dockerfile entrypoint to remove the validate command, which
allows the user to specify whether to validate or validate and count
the specified jobspec. Add a Flux broker option to suppress Flux error
output to avoid confusing LLM agents.